Added delete button for a day out. Working on functionality

+++ Controllers/DaysOutController.cs
@@ -196,6 +196,7 @@ namespace DadsDayApp.Controllers
         // to grab the id from the URL. It is then made available to us as the `id` argument to the method.
         //
         [HttpDelete("{id}")]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
         public async Task<IActionResult> DeleteDayOut(int id)
         {
             // Find this dayOut by looking for the specific id
@@ -205,6 +206,19 @@ namespace DadsDayApp.Controllers
                 // There wasn't a dayOut with that id so return a `404` not found
                 return NotFound();
             }
+            if (dayOut.UserId != GetCurrentUserId())
+            {
+                // Make a custom error response
+                var response = new
+                {
+                    status = 401,
+                    errors = new List<string>() { "Not Authorized" }
+                };
+
+                // Return our error with the custom response
+                return Unauthorized(response);
+            }
+
 
             // Tell the database we want to remove this record
             _context.DaysOut.Remove(dayOut);
